[ADD] animation ui barman

+++ PlatiniumProject/Assets/Scripts/Inputs&QTE/QTE/QTEHandler.cs
@@ -38,7 +38,7 @@ public class QTEHandler : MonoBehaviour, IIsControllable
     int _indexInListSequences = 0;
     bool _isSequenceComplete = false;
     int _durationHold = 0;
-    public float DurationValue { get => _currentQTESequence == null || _timingable == null ? 0 : _durationHold / (_currentQTESequence.DurationHold * _timingable.BeatDurationInMilliseconds); }
+    public float DurationValue { get => _currentQTESequence == null || _timingable == null ? 0 : _durationHold / (float)(_currentQTESequence.DurationHold * _timingable.BeatDurationInMilliseconds); }
     bool _hasHoldStarted = false;
     [SerializeField] float _holdDelayAcceptance = 0.5f;
     float _holdDelayAcceptanceTimer;
